Overflow Menu
At this point I have added the Overflow Menu,

To add an overflow menu (often represented by three dots) in the AppBar, you can use the DropdownMenu composable. This will allow you to display a list of options when the user taps the overflow icon.

Here’s how you can add an overflow menu to your TopAppBar:

Steps to Add an Overflow Menu:
Add an Overflow Icon: Use the Icons.Filled.MoreVert icon (the three-dot vertical icon) in the actions parameter of the TopAppBar.

Use a DropdownMenu: The DropdownMenu will appear when the user clicks the overflow icon. It contains menu items that the user can select.

Manage Menu State: You will need to use a remember state to control when the overflow menu is opened and closed.

Also, this is very Important
Use remember { mutableStateOf(false) } to store the state of whether the overflow menu is expanded or not.
Make sure you place this declaration inside your composable function.

Also add this import:

import androidx.compose.runtime.*
The code will work perfectly.